Move the scheduler to a new file.
